+++ addon/plugins/variable-plugin/variables/address.ts
@@ -1,6 +1,7 @@
 import {
   ADRES,
   EXT,
+  GEO,
 } from '@lblod/ember-rdfa-editor-lblod-plugins/utils/constants';
 import {
   createEmberNodeSpec,
@@ -26,13 +27,22 @@ import {
   mappingSpan,
   typeSpan,
 } from '../utils/dom-constructors';
+
+type Location = {
+  lat_WGS84: number;
+  long_WGS84: number;
+};
 export class Address {
   declare street: string;
   declare zipcode: string;
   declare municipality: string;
+  declare location: Location;
   housenumber?: string | null;
   constructor(
-    args: Pick<Address, 'street' | 'housenumber' | 'zipcode' | 'municipality'>,
+    args: Pick<
+      Address,
+      'street' | 'housenumber' | 'zipcode' | 'municipality' | 'location'
+    >,
   ) {
     Object.assign(this, args);
   }
@@ -51,7 +61,9 @@ export class Address {
         this.street === other.street &&
         this.housenumber === other.housenumber &&
         this.zipcode === other.zipcode &&
-        this.municipality === other.municipality
+        this.municipality === other.municipality &&
+        this.location.lat_WGS84 === other.location.lat_WGS84 &&
+        this.location.long_WGS84 === other.location.long_WGS84
       );
     } else {
       return false;
@@ -74,6 +86,7 @@ export class ResolvedAddress extends Address {
       | 'zipcode'
       | 'municipality'
       | 'addressRegisterId'
+      | 'location'
     >,
   ) {
     super(args);
@@ -87,6 +100,23 @@ export class ResolvedAddress extends Address {
     });
   }
 }
+
+const constructLocationNode = (location: Location) => {
+  return span(
+    {
+      property: ADRES('positie').full,
+      typeof: GEO('Point').full,
+    },
+    span({
+      property: GEO('lat').full,
+      content: location.lat_WGS84.toString(),
+    }),
+    span({
+      property: GEO('long').full,
+      content: location.long_WGS84.toString(),
+    }),
+  );
+};
 
 const constructAddressNode = (address?: Address | ResolvedAddress) => {
   if (address) {
@@ -130,10 +160,35 @@ const constructAddressNode = (address?: Address | ResolvedAddress) => {
         },
         address.municipality,
       ),
+      constructLocationNode(address.location),
     );
   } else {
     return contentSpan({}, 'Voeg adres in');
   }
+};
+
+const parseLocationNode = (locationNode: Element): Location | undefined => {
+  const lat_WGS84 = findChildWithRdfaAttribute(
+    locationNode,
+    'property',
+    GEO('lat'),
+  )?.getAttribute('content');
+  const long_WGS84 = findChildWithRdfaAttribute(
+    locationNode,
+    'property',
+    GEO('long'),
+  )?.getAttribute('content');
+  if (lat_WGS84 && long_WGS84) {
+    const lat_WGS84_number = parseFloat(lat_WGS84);
+    const long_WGS84_number = parseFloat(long_WGS84);
+    if (!isNaN(lat_WGS84_number) && !isNaN(long_WGS84_number)) {
+      return {
+        lat_WGS84: lat_WGS84_number,
+        long_WGS84: long_WGS84_number,
+      };
+    }
+  }
+  return;
 };
 
 const parseAddressNode = (addressNode: Element): Address | undefined => {
@@ -162,7 +217,13 @@ const parseAddressNode = (addressNode: Element): Address | undefined => {
     'property',
     ADRES('gemeentenaam'),
   )?.textContent;
-  if (street && zipcode && municipality) {
+  const locationNode = findChildWithRdfaAttribute(
+    addressNode,
+    'property',
+    ADRES('positie'),
+  );
+  const location = locationNode && parseLocationNode(locationNode);
+  if (street && zipcode && municipality && location) {
     if (addressRegisterId) {
       return new ResolvedAddress({
         addressRegisterId,
@@ -170,6 +231,7 @@ const parseAddressNode = (addressNode: Element): Address | undefined => {
         housenumber,
         zipcode,
         municipality,
+        location,
       });
     } else {
       return new Address({
@@ -177,6 +239,7 @@ const parseAddressNode = (addressNode: Element): Address | undefined => {
         housenumber,
         zipcode,
         municipality,
+        location,
       });
     }
   } else {